Squashed commit of the following:
commit 50b020a817f147fff3ef67aea680ab0ca6c798e0
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Apr 20 21:39:43 2026 +0200

    loading models from sourceforge if github fails

commit 0eb6cf346c2999ffb7a24f2e456cd42f6495fc96
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Apr 20 15:51:31 2026 +0200

    adjusted the .pyproj

commit 527c12141a8a710b635a9a030fb3f2723091a65b
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Apr 20 15:10:20 2026 +0200

    Update pipeline_worker.py

commit 0425dcc76d5d2cce70f20cde598e5938a8d0ac47
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Apr 20 15:02:11 2026 +0200

    exporting non-webm glb texture

commit 509256eefedd6000d7e1ac71e2a739d090ebb846
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Apr 20 12:06:34 2026 +0200

    adding api support

commit 3aa7cddf443721c91e7c11b58f96a72dc8ef931e
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Apr 6 23:34:26 2026 +0200

    app.py: decimation target limit reduced to 10k

commit 4a7e85173d55f7abea0d56e87ec285f700e3f790
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Apr 6 22:57:49 2026 +0200

    improved gradio + raised limit in sample_shape_slat_cascade 512->1024

commit 769a319b0bc649ee95a478e7d1b509c7bfff3661
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Apr 6 10:54:20 2026 +0200

    more tweaking install.py

commit c322bec112021e957ee01d99742b916c875965e6
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Apr 6 08:17:18 2026 +0200

    adjusted install script to download models

commit 830e3bf0176fcf8e16dfd2fa32670c2cba361e60
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Apr 6 06:23:09 2026 +0200

    setup.py references IgorAherne's nvdiffrec fork

commit a64d8fc667ce5ee5f416f271c0f09abcff787d0d
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Apr 6 05:53:20 2026 +0200

    corrected the nvdiffrec whl - avoiding garbage-init vals

commit e2c7997927c56a8f0b5c8f84b5beca79aa95a916
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Apr 6 05:11:09 2026 +0200

    flex_gemm using low chunk size only during Decoding

commit 90ef1798a895cb6064d8b54c94b19a3c3f0c8a4e
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Apr 6 04:21:50 2026 +0200

    Decoder VRAM 11.5→7.6GB: skip-to-CPU, fp16 dtype fixes, split mesh extraction

commit 8141d884fa23a863d83f9ac12fe54642ce057b31
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Apr 6 00:11:19 2026 +0200

    added solution

commit ac0979819c22a225732f6386d955d067cf13d961
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Wed Mar 25 15:36:22 2026 +0100

    cleanup

commit c8b6c9da95ea9d7d5af998adf605864981b3cb68
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Wed Mar 25 14:02:54 2026 +0100

    .

commit cad3154eac917c69690880da44433159deb7d346
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Wed Mar 25 13:30:46 2026 +0100

    rebuilt the cumesh + reduced vram during Extract GLB

commit ad8d94784bf50a7c9cc0a74721016dd40e85eaa2
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Wed Mar 25 02:57:04 2026 +0100

    .

commit 0f65dac138f6876ee397b08a1d9d59f99a4c8ab4
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Wed Mar 25 02:06:58 2026 +0100

    .

commit 00c20f477016c9d9021a1112bb2cfa8397a6acfa
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Wed Mar 25 01:22:11 2026 +0100

    1024 fits under 12GB

commit 1d6391b1b0eb9d95e472458f1fa319594aa1a01b
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Wed Mar 25 00:38:33 2026 +0100

    used custom sparse_conv3d_forward

commit 116a13f71c8ec85cb908b0593c7f1559665e4cdb
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Tue Mar 24 21:01:39 2026 +0100

    Update trellis2_image_to_3d.py

commit ddd2bd0fa0ce4774515b2bf55302ce7c5ecdc582
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Tue Mar 24 20:17:22 2026 +0100

    vram optimizations after `sampling texture SLat step`

commit 4b26aa778a85b30080ea0f5a4b1373974a2e1ff3
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Tue Mar 3 17:51:45 2026 +0100

    re-added profiling support

commit e417ac2efd7415f713b2f222b46f1593ef66d3e9
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Tue Mar 3 17:06:49 2026 +0100

    moving decoders back to gpu during GLB extraction

commit 0a009ed60627a849d8f2ae3d4c0b45709a680997
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Tue Mar 3 16:55:14 2026 +0100

    added pipeline_worker

commit 3e2674f06354452e827b6af824234ae9c104c5f7
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Tue Mar 3 12:40:52 2026 +0100

    unloading decoders

commit bb646e6667c227859acbd3ebf444ecf8889c7013
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Tue Mar 3 12:30:49 2026 +0100

    preventing GIL contention, by .sleep(0) inside flow_euler.py at every iter

commit 454660d1732b604d3a6d38c7207033ae828d4e85
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Tue Mar 3 03:12:39 2026 +0100

    optimizations

commit d2a0d85f4132d83401f396ed3ae0458aedf00525
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Mar 2 18:33:11 2026 +0100

    Update .gitignore

commit 401d313f0e882bc659698a020f197dbb55735ad0
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Mar 2 18:23:42 2026 +0100

    improved the profiling scripts

commit 90355e3c675ae29e0cecabc5b89e0b0efbf60bfe
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Mar 2 14:53:14 2026 +0100

    stop tracking .vs

commit 7917c50265331b2f9177da68db9db11c0eeded06
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Mar 2 14:47:10 2026 +0100

    updated profiling

commit db939ef8620242e929a0df0a70ea677b58835a64
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Mar 2 13:49:51 2026 +0100

    updated profiling script

commit f2bf2e61bf0c312f646c62ed868543e65f5e134d
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Thu Feb 12 05:57:08 2026 +0100

    flexgemm disabled triton

commit 21b5c2e3cd4aba0fc6187caa0c70fddec8036eb1
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Wed Feb 11 19:08:25 2026 +0100

    added profiling scripts

commit 63f98b8e2597ed73840a473647b243b25ecd9cc3
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Wed Feb 11 19:06:36 2026 +0100

    marked each file with its filepath

commit b6b97ea799f20b8557cbe85a9b9ec0ce8e1924f1
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Wed Feb 11 19:05:12 2026 +0100

    added more wheels

commit 3d0d3fca284374f76fe5dc7a41d90abb398b64e0
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Tue Dec 23 00:39:08 2025 +0100

    Create flex_gemm-0.0.1-cp311-cp311-win_amd64.whl

commit ec93ca229cb3bd6a08cb38050ab72a68c3573bd6
Author: DESKTOP-OCNK43A\Igor <igor.aherne.videos@gmail.com>
Date:   Mon Dec 22 22:57:00 2025 +0100

    added wheels

+++ trellis2/modules/utils.py
@@ -1,3 +1,4 @@
+# File: trellis2/modules/utils.py
 # trellis2/modules/utils.py
 import torch
 import torch.nn as nn
@@ -71,7 +72,7 @@ def manual_cast(tensor, dtype):
     Cast if autocast is not enabled.
     """
     if not torch.is_autocast_enabled():
-        return tensor.type(dtype)
+        return tensor.to(dtype)
     return tensor
